test: adding tests for the HTML statement
