+++ package.json
@@ -26,14 +26,14 @@
     "SDK"
   ],
   "dependencies": {
-    "@litexa/apl": "file:packages/litexa-apl",
-    "@litexa/assets-wav": "file:packages/litexa-assets-wav",
-    "@litexa/core": "file:packages/litexa",
-    "@litexa/deploy-aws": "file:packages/litexa-deploy-aws",
-    "@litexa/gadgets": "file:packages/litexa-gadgets",
-    "@litexa/html": "file:packages/litexa-html",
+    "@litexa/apl": "file:packages\\litexa-apl",
+    "@litexa/assets-wav": "file:packages\\litexa-assets-wav",
+    "@litexa/core": "file:packages\\litexa",
+    "@litexa/deploy-aws": "file:packages\\litexa-deploy-aws",
+    "@litexa/gadgets": "file:packages\\litexa-gadgets",
+    "@litexa/html": "file:packages\\litexa-html",
     "@litexa/integration-tests": "file:tests",
-    "@litexa/render-template": "file:packages/litexa-render-template"
+    "@litexa/render-template": "file:packages\\litexa-render-template"
   },
   "devDependencies": {
     "assert": "1.4.1",
