+++ packages/litexa-html/test/skill/litexa.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  name: "htmlTest"
+}
