+++ packages/litexa-html/test/skill/litexa/main.js
@@ -1,0 +1,15 @@
+function makeInitialData(appState) {
+  return {
+    appState: appState,
+    greeting: "Hello, player."
+  };
+}
+
+function makeInitialTransformers() { 
+  return [
+    {
+      transformer: 'textToSpeech',
+      inputPath: 'greeting',
+    }
+  ];
+}
