+++ packages/litexa-html/test/skill/litexa/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "litexa",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@litexa/html": "file:../../.."
+  }
+}
